requirements.txt
flask
requests
openai
wraps
load_dotenv
Flask-Cors

